Filter Gene Expression for Switching Genes

Create a reduced binary expression matrix for only the selected switching genes, binary_counts_matrix is from the sample DATA and gs_scorer_genes is from Atlas Data.

Usage

filter_gene_expression_for_switching_genes(binary_counts_matrix, gss_genes)

Arguments

binary_counts_matrix

a binary expression matrix from your sample.

gs_scorer_genes

Switching genes which are evenly distributed through pseudotime..

Value

a reduced binary expression matrix filtered to only include selected switching genes